requirements.txt
Django~=4.2.6
djangorestframework~=3.14.0
django-cors-headers~=4.3.0
drf-yasg~=1.21.7
black~=23.10.1
pylint~=3.0.2
flake8~=6.1.0
python-dotenv~=1.0.0
numpy~=1.24.4
backports.zoneinfo;python_version<"3.9"
Pillow~=10.1.0